repo structure: move headers to new include/microdns folder
With that change, libmicrodns can be used as a meson subproject
without requiring the parent project to use a different include
path.

This also makes the examples more readily copypastable, as the
include is switched from `#include "microdns.h"` to
`#include <microdns/microdns.h`, which is how external compilation
units are expected to perform the inclusion.

In addition, this also removes the unneeded compat.h inclusion from
the announce example, this has been tested and is known to compile
on Windows with mingw / MSYS2.
