+++ compat/inet.c
@@ -25,7 +25,7 @@
 #include <string.h>
 
 #include "compat.h"
-#include "utils.h"
+#include "microdns/utils.h"
 
 #if defined (_WIN32) && !defined(HAVE_INET_NTOP)
 const char *
