+++ examples/main.c
@@ -20,7 +20,7 @@
 #include <unistd.h>
 #include <signal.h>
 
-#include "microdns.h"
+#include <microdns/microdns.h>
 
 volatile sig_atomic_t sigflag = 0;
