+++ src/mdns.c
@@ -35,8 +35,8 @@
 #include <stdint.h>
 
 #include "compat.h"
-#include "utils.h"
-#include "microdns.h"
+#include "microdns/utils.h"
+#include "microdns/microdns.h"
 
 #define MDNS_PKT_MAXSZ 4096 // read/write buffer size
